requirements.txt
requests
boto3
awscli
#psycopg2 #installed through libs
pytest
testing.postgresql
gtfs-realtime-bindings